+++ tests/test_printers.py
@@ -34,18 +34,3 @@ def test_find_dict_composite_key_value(obj, path, res):
 def test_find_dict_composite_key_value_fails():
     with pytest.raises(AttributeError):
         printers.find_dict_composite_key_value({"a": "b"}, "a.b")
-
-
-@pytest.mark.parametrize(
-    "asset,output_format,printer_cls",
-    [
-        ("compute_plan", "pretty", printers.AssetPrinter),
-        ("compute_plan", "json", printers.JsonPrinter),
-        ("compute_plan", "yaml", printers.YamlPrinter),
-        ("foo", "pretty", printers.JsonPrinter),
-        ("foo", "json", printers.JsonPrinter),
-        ("foo", "yaml", printers.YamlPrinter),
-    ],
-)
-def test_get_asset_printer(asset, output_format, printer_cls):
-    assert isinstance(printers.get_asset_printer(asset, output_format), printer_cls)